Move VesselReportingList/ReportingCard to parent in Frontend
